New package: bootstrap-iconds
The Bootstrap Icons library is a free, high-quality, open-source icon
library with over 1,600 icons. It can be included in all kind of
packages and it does not depend on Bootstrap at all. Include them
anyway you like—SVGs, SVG sprite, or web fonts.

The package support installation from CDN and local installation,
switchable via the site-wide administration link for the package in
/acs-admin