Fix logic to retrieve the Boostrap Icons CSS path:
the situation is even more slippery than thought, as it seems the new "font" path was discontinued for versions 1.11.0 and 1.11.1, then reintroduced in version 1.11.2. We can only speculate what logic will follow the next version, so for now we assume the new path will stay and just skip the two outlier versions.